CARDS-1669: Add support for exporting data as csv Move classes in the right place
