+++ modules/utils/src/main/java/io/uhndata/cards/serialize/CSVString.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.uhndata.cards.dataentry.internal.serialize;
+package io.uhndata.cards.serialize;
 
 /**
  * Shell class for CSV adaptor.
